requirements.txt
bottle
bottle-sqlalchemy
pytest
requests